+++ .github/workflows/test-run-script.yml
@@ -1,9 +1,8 @@
-on:
-  push
+on: push
 
 jobs:
   test_run_script_job:
     uses: ./.github/workflows/run-script.yml
     with:
-      script_name_with_args: parse-review-issues
+      script_name_with_args: update_contributors
     secrets: inherit
